fixed issues with conflicting auth and setup wrappers as well as the google login create user issue
